requirements.txt
numpy
torch
torchvision
opencv-python
Pillow
scikit-image
sys
PyQt5
math
matplotlib